+++ packages/dodo-ui-date/src/lib/components/DatePicker/DatePickerInput/DatePickerInput.svelte
@@ -122,6 +122,8 @@
   class={classes.join(' ')}
   focused={forcedFocused || isFocused}
   bind:ref
+  affixSpacingAfter={triggerPlacement === 'after' ? 'tight' : 'default'}
+  affixSpacingBefore={triggerPlacement === 'before' ? 'tight' : 'default'}
   {...restProps}
 >
   <DatePicker.Input